requirements.txt
supabase
beautifulsoup4
requests
python-dotenv
mcp[cli]
openai
psycopg2-binary
